First draft of the code
This code was written for a gas sensor, measures the amount of smoke and other harmful gases in the air. If it reaches some predefined tresholds the user is alarmed by led diodes ( green - be aware, yellow - the concentration of the gas is increasing, red - leave the place, the health risk is high) and also by a buzzer.